add confirmation modal

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { TitleCasePipe } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -17,7 +18,7 @@ export class MessagesComponent implements OnInit {
 	pageSize = 6;
 	loading = false;
 
-	constructor(private messageSerivce: MessageService) {}
+	constructor(private messageSerivce: MessageService, private confirmService: ConfirmService) {}
 
 	ngOnInit(): void {
 		this.loadMessages();
@@ -43,11 +44,15 @@ export class MessagesComponent implements OnInit {
 	}
 
 	deleteMessage(id: number) {
-		this.messageSerivce.deleteMessage(id).subscribe(() => {
-			this.messages.splice(
-				this.messages.findIndex((m) => m.id === id),
-				1
-			);
+		this.confirmService.confirm('Confirm delete message', 'This cannot be undone').subscribe((result) => {
+			if (result) {
+				this.messageSerivce.deleteMessage(id).subscribe(() => {
+					this.messages.splice(
+						this.messages.findIndex((m) => m.id === id),
+						1
+					);
+				});
+			}
 		});
 	}
 }
